feat(streamback): support multiple topics in listeners
This update modifies the `Listener` class in streamback to support multiple topics per listener. The `topic` attribute is replaced with `topics` which is a list of topics that the listener is subscribed to. This change also affects the associated classes and methods that use the `Listener` class.

In addition, the maximum partition fetch bytes for the Kafka consumer configuration has been increased from 16 to 128. The Kafka consumer now consumes 100 messages at a time and the `Stream` class now has a `closed` attribute to indicate if the stream is closed. The `read_stream` method of the Kafka stream now runs until the stream is closed.