feat: update version to 0.5.1 and enhance KafkaStream configuration
This commit bumps the version in setup.py to 0.5.1. It also updates the Kafka service configuration in docker-compose.yml to use a specific Confluent Kafka image and adds new services for futures-consumer and futures-publisher. Additionally, the Streamback class is modified to support a group instance ID for Kafka consumers, improving process management and listener functionality.